chore(cmd): migrate keys command to cobra style
